refactor: reorganize project structure and split Program.cs
Directory Restructure:
- Configuration/   : MyOption, BBDownConfigParser, ServeRequestOptions
- Infrastructure/  : BBDownApiServer, BBDownAria2c, BBDownDownloadUtil,
                     BBDownLoginUtil, BBDownMuxer, ConsoleQRCode, ProgressBar
- Models/          : BBDownEnums
- Utilities/       : BBDownUtil
- Application/     : Partial Program class split by responsibility

Program.cs Split (1026 → 144 lines):
- Program.cs        : Entry point, Main, StartServer, DoWorkAsync
- Workflow.cs       : SetUpWork, GetVideoInfoAsync
- Download.cs       : DownloadPagesAsync, DownloadPageAsync, DecryptDrmAsync
- Options.cs        : CLI option parsing and setup
- Pages.cs          : Page selection, PCDN handling
- Display.cs        : Track display, manual selection, DownloadTrackAsync
- Archive.cs        : Save/check aid archive file
- PathHelper.cs     : FormatSavePath, FindTool, InfoRegex
- TrackSort.cs      : SortTracks
- Decrypt.cs        : DRM decryption helpers
- ToolFinder.cs     : External binary resolution
- Regex.cs          : Generated Regex patterns

Removed: Program.Methods.cs (535 lines, fully absorbed into new files)

Build: 0 Warning(s), 0 Error(s)
