+++ BBDown/Infrastructure/ConsoleQRCode.cs
