+++ BBDown/Program.cs
@@ -119,658 +119,6 @@ partial class Program
         server.Run(string.IsNullOrEmpty(listenUrl) ? defaultListenUrl : listenUrl);
     }
 
-    public static (Dictionary<string, byte> encodingPriority, Dictionary<string, int> dfnPriority, string? firstEncoding,
-        bool downloadDanmaku, BBDownDanmakuFormat[] downloadDanmakuFormats, string input, string savePathFormat, string lang, string aidOri, int delay)
-        SetUpWork(MyOption myOption)
-    {
-        //处理废弃选项
-        HandleDeprecatedOptions(myOption);
-
-        //处理冲突选项
-        HandleConflictingOptions(myOption);
-
-        //寻找并设置所需的二进制文件路径
-        FindBinaries(myOption);
-
-        //切换工作目录
-        ChangeWorkingDir(myOption);
-
-        //解析优先级
-        var encodingPriority = ParseEncodingPriority(myOption, out var firstEncoding);
-        var dfnPriority = ParseDfnPriority(myOption);
-
-        //优先使用用户设置的UA
-        HTTPUtil.UserAgent = string.IsNullOrEmpty(myOption.UserAgent) ? HTTPUtil.UserAgent : myOption.UserAgent;
-
-        bool downloadDanmaku = myOption.DownloadDanmaku || myOption.DanmakuOnly;
-        BBDownDanmakuFormat[] downloadDanmakuFormats = ParseDownloadDanmakuFormats(myOption);
-
-        string input = myOption.Url;
-        string savePathFormat = myOption.FilePattern;
-        string lang = myOption.Language;
-        string aidOri = ""; //原始aid
-        int delay = myOption.DelayPerPage;
-        Config.DEBUG_LOG = myOption.Debug;
-        Config.SKIP_SSL_CHECK = myOption.Insecure;
-        Config.HOST = myOption.Host;
-        Config.EPHOST = myOption.EpHost;
-        Config.TVHOST = myOption.TvHost;
-        Config.AREA = myOption.Area;
-        Config.COOKIE = myOption.Cookie;
-        Config.TOKEN = myOption.AccessToken.Replace("access_token=", "");
-
-        LogDebug("AppDirectory: {0}", APP_DIR);
-        if (Config.DEBUG_LOG)
-        {
-            var savedCookie = myOption.Cookie;
-            var savedToken = myOption.AccessToken;
-            myOption.Cookie = string.IsNullOrEmpty(savedCookie) ? "" : "***";
-            myOption.AccessToken = string.IsNullOrEmpty(savedToken) ? "" : "***";
-            LogDebug("运行参数：{0}", JsonSerializer.Serialize(myOption, MyOptionJsonContext.Default.MyOption));
-            myOption.Cookie = savedCookie;
-            myOption.AccessToken = savedToken;
-        }
-        return (encodingPriority, dfnPriority, firstEncoding, downloadDanmaku, downloadDanmakuFormats, input, savePathFormat, lang, aidOri, delay);
-    }
-
-    public static async Task<(string fetchedAid, VInfo vInfo, string apiType)> GetVideoInfoAsync(MyOption myOption, string aidOri, string input)
-    {
-        // 加载认证信息
-        LoadCredentials(myOption);
-
-        // 检测是否登录了账号
-        if (myOption is { UseIntlApi: false, UseTvApi: false } && Config.AREA == "")
-        {
-            Log("检测账号登录...");
-            if (!await CheckLogin(Config.COOKIE))
-            {
-                LogWarn("========================================");
-                LogWarn("  你尚未登录B站账号！");
-                LogWarn("  未登录状态下仅能下载6分钟试看片段。");
-                LogWarn("  请运行 BBDown login 扫码登录以获取完整视频。");
-                LogWarn("========================================");
-            }
-        }
-
-        Log("获取aid...");
-        aidOri = await GetAvIdAsync(input);
-        Log($"获取aid结束: {aidOri}");
-
-        if (string.IsNullOrEmpty(aidOri))
-        {
-            throw new Exception("输入有误");
-        }
-
-        Log("获取视频信息...");
-        IFetcher fetcher = FetcherFactory.CreateFetcher(aidOri, myOption.UseIntlApi);
-        VInfo? vInfo = null;
-
-        // 只输入 EP/SS 时优先按番剧查找，如果找不到则尝试按课程查找
-        try
-        {
-            vInfo = await fetcher.FetchAsync(aidOri);
-        }
-        catch (Exception e) when (e is KeyNotFoundException or InvalidOperationException)
-        {
-            // B站返回非番剧JSON结构（可能是课程），尝试按课程查找
-            if (aidOri.StartsWith("cheese:")) throw; // 已经按课程查找过，不再重复尝试
-
-            LogWarn("未找到此 EP/SS 对应番剧信息, 正在尝试按课程查找。");
-
-            aidOri = aidOri.Replace("ep", "cheese");
-            Log("新的 aid: " + aidOri);
-
-            if (string.IsNullOrEmpty(aidOri))
-            {
-                throw new Exception("输入有误");
-            }
-
-            Log("获取视频信息...");
-            fetcher = FetcherFactory.CreateFetcher(aidOri, myOption.UseIntlApi);
-            vInfo = await fetcher.FetchAsync(aidOri);
-        }
-
-        string title = vInfo.Title;
-        long pubTime = vInfo.PubTime;
-        LogColor("视频标题: " + title);
-        if (pubTime != 0)
-        {
-            Log("发布时间: " + FormatTimeStamp(pubTime, "yyyy-MM-dd HH:mm:ss zzz"));
-        }
-        var bvid = vInfo.PagesInfo.FirstOrDefault()?.bvid;
-        if (!string.IsNullOrEmpty(bvid) && !myOption.UseIntlApi)
-        {
-            Log($"视频URL: https://www.bilibili.com/video/{bvid}/");
-        }
-        var mid = vInfo.PagesInfo.FirstOrDefault(p => !string.IsNullOrEmpty(p.ownerMid))?.ownerMid;
-        if (!string.IsNullOrEmpty(mid))
-        {
-            Log($"UP主页: https://space.bilibili.com/{mid}");
-        }
-
-        if (vInfo.IsSteinGate && myOption.UseTvApi)
-        {
-            Log("视频为互动视频，暂时不支持tv下载，修改为默认下载");
-            myOption.UseTvApi = false;
-        }
-        string apiType = myOption.UseTvApi ? "TV" : (myOption.UseAppApi ? "APP" : (myOption.UseIntlApi ? "INTL" : "WEB"));
-
-        //打印分P信息
-        List<Page> pagesInfo = vInfo.PagesInfo;
-        bool more = false;
-        foreach (Page p in pagesInfo)
-        {
-            if (!myOption.ShowAll)
-            {
-                if (more && p.index != pagesInfo.Count) continue;
-                if (!more && p.index > 5)
-                {
-                    Log("......");
-                    more = true;
-                    continue;
-                }
-            }
-
-            Log($"P{p.index}: [{p.cid}] [{p.title}] [{FormatTime(p.dur)}]");
-        }
-        return (aidOri, vInfo, apiType);
-    }
-
-    public static async Task DownloadPagesAsync(MyOption myOption, VInfo vInfo, Dictionary<string, byte> encodingPriority, Dictionary<string, int> dfnPriority,
-        string? firstEncoding, bool downloadDanmaku, BBDownDanmakuFormat[] downloadDanmakuFormats, string input, string savePathFormat, string lang, string aidOri, int delay, string apiType, DownloadTask? relatedTask = null)
-    {
-        List<Page> pagesInfo = vInfo.PagesInfo;
-        bool bangumi = vInfo.IsBangumi;
-        bool cheese = vInfo.IsCheese;
-        //获取已选择的分P列表
-        List<string>? selectedPages = GetSelectedPages(myOption, vInfo, input);
-
-        Log($"共计 {pagesInfo.Count} 个分P, 已选择：" + (selectedPages == null ? "ALL" : string.Join(",", selectedPages)));
-        var pagesCount = pagesInfo.Count;
-
-        //过滤不需要的分P
-        if (selectedPages != null)
-        {
-            pagesInfo = pagesInfo.Where(p => selectedPages.Contains(p.index.ToString())).ToList();
-        }
-
-        // 根据p数选择存储路径
-        savePathFormat = string.IsNullOrEmpty(myOption.FilePattern) ? SinglePageDefaultSavePath : myOption.FilePattern;
-        // 1. 多P; 2. 只有1P, 但是是番剧, 尚未完结时 按照多P处理
-        if (pagesCount > 1 || (bangumi && !vInfo.IsBangumiEnd))
-        {
-            savePathFormat = string.IsNullOrEmpty(myOption.MultiFilePattern) ? MultiPageDefaultSavePath : myOption.MultiFilePattern;
-        }
-
-        foreach (Page p in pagesInfo)
-        {
-            if (pagesInfo.Count > 1 && delay > 0)
-            {
-                Log($"停顿{delay}秒...");
-                await Task.Delay(delay * 1000);
-            }
-            Log($"开始解析P{p.index}: {p.aid}... ({pagesInfo.IndexOf(p) + 1} of {pagesInfo.Count})");
-
-            if (myOption.SaveArchivesToFile)
-            {
-                if (CheckAidFromFile(p.aid))
-                {
-
-                    Log($"aid: {p.aid}已下载过, 跳过下载...");
-                    continue;
-                }
-            }
-
-            await DownloadPageAsync(p, myOption, vInfo, pagesInfo, encodingPriority, dfnPriority, firstEncoding,
-                downloadDanmaku, downloadDanmakuFormats, input, savePathFormat, lang, aidOri, apiType, relatedTask);
-
-            if (myOption.SaveArchivesToFile)
-            {
-                SaveAidToFile(p.aid);
-            }
-        }
-
-        Log("任务完成");
-    }
-
-    private static async Task DownloadPageAsync(Page p, MyOption myOption, VInfo vInfo, List<Page> selectedPagesInfo, Dictionary<string, byte> encodingPriority, Dictionary<string, int> dfnPriority,
-        string? firstEncoding, bool downloadDanmaku, BBDownDanmakuFormat[] downloadDanmakuFormats, string input, string savePathFormat, string lang, string aidOri, string apiType, DownloadTask? relatedTask = null)
-    {
-        string desc = string.IsNullOrEmpty(p.desc) ? vInfo.Desc : p.desc;
-        bool bangumi = vInfo.IsBangumi;
-        var pagesCount = selectedPagesInfo.Count;
-        List<Subtitle> subtitleInfo = [];
-        string title = vInfo.Title;
-        string pic = vInfo.Pic;
-        long pubTime = vInfo.PubTime;
-        bool selected = false; //用户是否已经手动选择过了轨道
-        int retryCount = 0;
-        while (retryCount < 3)
-        {
-        try
-        {
-            LogDebug("尝试获取章节信息...");
-            p.points = await FetchPointsAsync(p.cid, p.aid);
-
-            string videoPath = $"{p.aid}/{p.aid}.P{p.index}.{p.cid}.mp4";
-            string audioPath = $"{p.aid}/{p.aid}.P{p.index}.{p.cid}.m4a";
-            var coverPath = $"{p.aid}/{p.aid}.jpg";
-
-            //处理文件夹以.结尾导致的异常情况
-            if (title.EndsWith('.')) title += "_fix";
-            //处理文件夹以.开头导致的异常情况
-            if (title.StartsWith('.')) title = "_" + title;
-
-            //处理封面&&字幕
-            if (!myOption.OnlyShowInfo)
-            {
-                if (!Directory.Exists(p.aid))
-                {
-                    Directory.CreateDirectory(p.aid);
-                }
-                if (!myOption.SkipCover && !myOption.SubOnly && !File.Exists(coverPath) && !myOption.DanmakuOnly && !myOption.CoverOnly)
-                {
-                    await DownloadFileAsync(pic == "" ? p.cover! : pic, coverPath, new DownloadConfig());
-                }
-
-                if (!myOption.SkipSubtitle && !myOption.DanmakuOnly && !myOption.CoverOnly)
-                {
-                    LogDebug("获取字幕...");
-                    subtitleInfo = await SubUtil.GetSubtitlesAsync(p.aid, p.cid, p.epid, p.index, myOption.UseIntlApi);
-                    if (myOption.SkipAi && subtitleInfo.Any())
-                    {
-                        Log($"跳过下载AI字幕");
-                        subtitleInfo = subtitleInfo.Where(s => !s.lan.StartsWith("ai-")).ToList();
-                    }
-                    foreach (Subtitle s in subtitleInfo)
-                    {
-                        Log($"下载字幕 {s.lan} => {SubUtil.GetSubtitleCode(s.lan).Item2}...");
-                        LogDebug("下载：{0}", s.url);
-                        await SubUtil.SaveSubtitleAsync(s.url, s.path);
-                        if (myOption.SubOnly && File.Exists(s.path) && File.ReadAllText(s.path) != "")
-                        {
-                            var _outSubPath = FormatSavePath(savePathFormat, title, null, null, p, pagesCount, apiType, pubTime);
-                            var dir = Path.GetDirectoryName(_outSubPath);
-                            if (!string.IsNullOrEmpty(dir) && !Directory.Exists(dir))
-                                Directory.CreateDirectory(dir);
-                            _outSubPath = Path.ChangeExtension(_outSubPath, $".{s.lan}.srt");
-                            File.Move(s.path, _outSubPath, true);
-                        }
-                    }
-                }
-
-                if (myOption.SubOnly)
-                {
-                    if (Directory.Exists(p.aid) && Directory.GetFiles(p.aid).Length == 0) Directory.Delete(p.aid, true);
-                    return;
-                }
-            }
-
-            //调用解析
-            ParsedResult parsedResult = await ExtractTracksAsync(aidOri, p.aid, p.cid, p.epid, myOption.UseTvApi, myOption.UseIntlApi, myOption.UseAppApi, firstEncoding!, myOption.DecryptDrm);
-            List<AudioMaterial> audioMaterial = [];
-            if (!p.points.Any())
-            {
-                p.points = parsedResult.ExtraPoints;
-            }
-
-            if (Config.DEBUG_LOG)
-            {
-                var debugFile = $"debug_{DateTime.Now:yyyyMMddHHmmssfff}.json";
-                File.WriteAllText(debugFile, parsedResult.WebJsonString);
-                // 限制 debug 文件数量，保留最近 20 个
-                var debugFiles = Directory.GetFiles(".", "debug_*.json").Order().ToArray();
-                for (int i = 0; i < debugFiles.Length - 20; i++)
-                    File.Delete(debugFiles[i]);
-            }
-
-            var savePath = "";
-
-            var downloadConfig = new DownloadConfig()
-            {
-                UseAria2c = myOption.UseAria2c,
-                Aria2cArgs = myOption.Aria2cArgs,
-                ForceHttp = myOption.ForceHttp,
-                MultiThread = myOption.MultiThread,
-                RelatedTask = relatedTask,
-            };
-
-            //此处代码简直灾难, 后续优化吧
-            if ((parsedResult.VideoTracks.Any() || parsedResult.AudioTracks.Any()) && !parsedResult.Clips.Any())   //dash
-            {
-                if (parsedResult.VideoTracks.Count == 0)
-                {
-                    LogWarn("没有找到符合要求的视频流");
-                    if (myOption.VideoOnly) return;
-                }
-                if (parsedResult.AudioTracks.Count == 0)
-                {
-                    LogWarn("没有找到符合要求的音频流");
-                    if (myOption.AudioOnly) return;
-                }
-
-                if (myOption.AudioOnly)
-                {
-                    parsedResult.VideoTracks.Clear();
-                }
-                if (myOption.VideoOnly)
-                {
-                    parsedResult.AudioTracks.Clear();
-                    parsedResult.BackgroundAudioTracks.Clear();
-                    parsedResult.RoleAudioList.Clear();
-                }
-
-                //排序
-                parsedResult.VideoTracks = SortTracks(parsedResult.VideoTracks, dfnPriority, encodingPriority, myOption.VideoAscending);
-                parsedResult.AudioTracks = SortTracks(parsedResult.AudioTracks, encodingPriority, myOption.AudioAscending);
-                parsedResult.BackgroundAudioTracks = SortTracks(parsedResult.BackgroundAudioTracks, encodingPriority, myOption.AudioAscending);
-                foreach (var role in parsedResult.RoleAudioList)
-                {
-                    role.audio = SortTracks(role.audio, encodingPriority, myOption.AudioAscending);
-                }
-
-                //打印轨道信息
-                if (!myOption.HideStreams)
-                {
-                    PrintAllTracksInfo(parsedResult, p.dur, myOption.OnlyShowInfo);
-                }
-
-                //仅展示 跳过下载
-                if (myOption.OnlyShowInfo)
-                {
-                    return;
-                }
-
-                int vIndex = 0; //用户手动选择的视频序号
-                int aIndex = 0; //用户手动选择的音频序号
-
-                //选择轨道
-                if (myOption.Interactive && !selected)
-                {
-                    SelectTrackManually(parsedResult, ref vIndex, ref aIndex);
-                    selected = true;
-                }
-
-                Video? selectedVideo = parsedResult.VideoTracks.ElementAtOrDefault(vIndex);
-                Audio? selectedAudio = parsedResult.AudioTracks.ElementAtOrDefault(aIndex);
-                Audio? selectedBackgroundAudio = parsedResult.BackgroundAudioTracks.ElementAtOrDefault(aIndex);
-
-                LogDebug("Format Before: " + savePathFormat);
-                savePath = FormatSavePath(savePathFormat, title, selectedVideo, selectedAudio, p, pagesCount, apiType, pubTime);
-                LogDebug("Format After: " + savePath);
-
-                if (downloadDanmaku)
-                {
-                    var danmakuXmlPath = Path.ChangeExtension(savePath, ".xml");
-                    var danmakuAssPath = Path.ChangeExtension(savePath, ".ass");
-                    Log("正在下载弹幕Xml文件");
-                    var danmakuUrl = $"https://comment.bilibili.com/{p.cid}.xml";
-                    await DownloadFileAsync(danmakuUrl, danmakuXmlPath, downloadConfig);
-                    var danmakus = DanmakuUtil.ParseXml(danmakuXmlPath);
-                    if (danmakus == null)
-                    {
-                        Log("弹幕Xml解析失败, 删除Xml...");
-                        File.Delete(danmakuXmlPath);
-                    }
-                    else if (danmakus.Length == 0)
-                    {
-                        Log("当前视频没有弹幕, 删除Xml...");
-                        File.Delete(danmakuXmlPath);
-                    }
-                    else if (downloadDanmakuFormats.Contains(BBDownDanmakuFormat.Ass))
-                    {
-                        Log("正在保存弹幕Ass文件...");
-                        await DanmakuUtil.SaveAsAssAsync(danmakus, danmakuAssPath);
-                    }
-
-                    // delete xml if possible
-                    if (!downloadDanmakuFormats.Contains(BBDownDanmakuFormat.Xml) && File.Exists(danmakuXmlPath)) 
-                    {
-                        File.Delete(danmakuXmlPath);
-                    }
-
-                    if (myOption.DanmakuOnly)
-                    {
-                        if (Directory.Exists(p.aid))
-                        {
-                            Directory.Delete(p.aid);
-                        }
-                        return;
-                    }
-                }
-
-                if (myOption.CoverOnly)
-                {
-                    var coverUrl = pic == "" ? p.cover! : pic;
-                    var newCoverPath = Path.ChangeExtension(savePath, Path.GetExtension(coverUrl));
-                    await DownloadFileAsync(coverUrl, newCoverPath, downloadConfig);
-                    if (Directory.Exists(p.aid) && Directory.GetFiles(p.aid).Length == 0) Directory.Delete(p.aid, true);
-                    relatedTask?.SavePaths.Add(newCoverPath);
-                }
-
-                Log($"已选择的流:");
-                PrintSelectedTrackInfo(selectedVideo, selectedAudio, p.dur);
-
-                //用户开启了强制替换
-                if (myOption.ForceReplaceHost && string.IsNullOrEmpty(myOption.UposHost))
-                {
-                    myOption.UposHost = BACKUP_HOST;
-                }
-
-                //处理PCDN
-                HandlePcdn(myOption, selectedVideo, selectedAudio);
-
-                if (!myOption.OnlyShowInfo && File.Exists(savePath) && new FileInfo(savePath).Length != 0)
-                {
-                    Log($"{savePath}已存在, 跳过下载...");
-                    relatedTask?.SavePaths.Add(savePath);
-                    File.Delete(coverPath);
-                    if (Directory.Exists(p.aid) && Directory.GetFiles(p.aid).Length == 0)
-                    {
-                        Directory.Delete(p.aid, true);
-                    }
-                    return;
-                }
-
-                if (selectedVideo != null)
-                {
-                    //杜比视界, 若ffmpeg版本小于5.0, 使用mp4box封装
-                    if (selectedVideo.dfn == Config.qualitys["126"] && !myOption.UseMP4box && !CheckFFmpegDOVI())
-                    {
-                        LogWarn($"检测到杜比视界清晰度且您的ffmpeg版本小于5.0,将使用mp4box混流...");
-                        myOption.UseMP4box = true;
-                    }
-                    Log($"开始下载P{p.index}视频...");
-                    await DownloadTrackAsync(selectedVideo.baseUrl, videoPath, downloadConfig, video: true);
-                }
-
-                if (selectedAudio != null)
-                {
-                    Log($"开始下载P{p.index}音频...");
-                    await DownloadTrackAsync(selectedAudio.baseUrl, audioPath, downloadConfig, video: false);
-                }
-
-                if (selectedBackgroundAudio != null)
-                {
-                    var backgroundPath = $"{p.aid}/{p.aid}.{p.cid}.P{p.index}.back_ground.m4a";
-                    Log($"开始下载P{p.index}背景配音...");
-                    await DownloadTrackAsync(selectedBackgroundAudio.baseUrl, backgroundPath, downloadConfig, video: false);
-                    audioMaterial.Add(new AudioMaterial("背景音频", "", backgroundPath));
-                }
-
-                if (parsedResult.RoleAudioList.Any())
-                {
-                    foreach (var role in parsedResult.RoleAudioList)
-                    {
-                        Log($"开始下载P{p.index}配音[{role.title}]...");
-                        await DownloadTrackAsync(role.audio[aIndex].baseUrl, role.path, downloadConfig, video: false);
-                        audioMaterial.Add(new AudioMaterial(role));
-                    }
-                }
-
-                Log($"下载P{p.index}完毕");
-
-                if (parsedResult.IsDrm && myOption.DecryptDrm && (!string.IsNullOrEmpty(parsedResult.KidHex) || !string.IsNullOrEmpty(parsedResult.PsshBase64)))
-                {
-                    await DecryptDrmAsync(parsedResult, videoPath, audioPath, myOption);
-                }
-
-                if (!parsedResult.VideoTracks.Any()) videoPath = "";
-                if (!parsedResult.AudioTracks.Any()) audioPath = "";
-                if (myOption.SkipMux) return;
-                Log($"开始合并音视频{(subtitleInfo.Any() ? "和字幕" : "")}...");
-                if (myOption.AudioOnly)
-                    savePath = savePath[..^4] + ".m4a";
-
-                var isHevc = selectedVideo?.codecs == "HEVC";
-                int code = BBDownMuxer.MuxAV(myOption.UseMP4box, p.bvid, videoPath, audioPath, audioMaterial, savePath,
-                    desc,
-                    title,
-                    p.ownerName ?? "",
-                    (pagesCount > 1 || (bangumi && !vInfo.IsBangumiEnd)) ? p.title : "",
-                    File.Exists(coverPath) ? coverPath : "",
-                    lang,
-                    subtitleInfo, myOption.AudioOnly, myOption.VideoOnly, p.points, p.pubTime, myOption.SimplyMux, isHevc);
-                if (code != 0 || !File.Exists(savePath) || new FileInfo(savePath).Length == 0)
-                {
-                    LogError("合并失败"); return;
-                }
-                Log("清理临时文件...");
-                await Task.Delay(200);
-                if (parsedResult.VideoTracks.Any()) File.Delete(videoPath);
-                if (parsedResult.AudioTracks.Any()) File.Delete(audioPath);
-                if (p.points.Any()) File.Delete(Path.Combine(Path.GetDirectoryName(string.IsNullOrEmpty(videoPath) ? audioPath : videoPath)!, "chapters"));
-                foreach (var s in subtitleInfo) File.Delete(s.path);
-                foreach (var a in audioMaterial) File.Delete(a.path);
-                if (selectedPagesInfo.Count == 1 || p.index == selectedPagesInfo.Last().index || p.aid != selectedPagesInfo.Last().aid)
-                    File.Delete(coverPath);
-                if (Directory.Exists(p.aid) && Directory.GetFiles(p.aid).Length == 0) Directory.Delete(p.aid, true);
-            }
-            else if (parsedResult.Clips.Any() && parsedResult.Dfns.Any())   //flv
-            {
-                if (myOption.DecryptDrm)
-                {
-                    LogError("此视频需要大会员登录才能获取完整DRM内容。");
-                    LogError($"请先运行: BBDown login  或使用 --cookie 参数");
-                    return;
-                }
-                var clips = parsedResult.Clips;
-                var dfns = parsedResult.Dfns;
-
-                int vIndex = 0;
-                if (myOption.Interactive && !selected)
-                {
-                    int i = 0;
-                    dfns.ForEach(key => LogColor($"{i++}.{Config.qualitys[key]}"));
-                    Log("请选择最想要的清晰度(输入序号): ", false);
-                    Console.ForegroundColor = ConsoleColor.Cyan;
-                    vIndex = ReadIntSafe();
-                    if (vIndex > dfns.Count || vIndex < 0) vIndex = 0;
-                    Console.ResetColor();
-                    //重新解析
-                    parsedResult = await ExtractTracksAsync(aidOri, p.aid, p.cid, p.epid, myOption.UseTvApi, myOption.UseIntlApi, myOption.UseAppApi, firstEncoding!, myOption.DecryptDrm, dfns[vIndex]);
-                    if (!p.points.Any()) p.points = parsedResult.ExtraPoints;
-                    selected = true;
-                    vIndex = 0; // 重新解析后第一个轨道即为所选清晰度
-                }
-                //排序
-                parsedResult.VideoTracks = SortTracks(parsedResult.VideoTracks, dfnPriority, encodingPriority, myOption.VideoAscending);
-
-                Log($"共计{parsedResult.VideoTracks.Count}条流(共有{clips.Count}个分段).");
-                int index = 0;
-                foreach (var v in parsedResult.VideoTracks)
-                {
-                    LogColor($"{index++}. [{v.dfn}] [{v.res}] [{v.codecs}] [{v.fps}] [~{v.size / 1024 / v.dur * 8:00} kbps] [{FormatFileSize(v.size)}]".Replace("[] ", ""), false);
-                    if (myOption.OnlyShowInfo)
-                    {
-                        clips.ForEach(Console.WriteLine);
-                    }
-                }
-                if (myOption.OnlyShowInfo) return;
-                savePath = FormatSavePath(savePathFormat, title, parsedResult.VideoTracks.ElementAtOrDefault(vIndex), null, p, pagesCount, apiType, pubTime);
-                if (File.Exists(savePath) && new FileInfo(savePath).Length != 0)
-                {
-                    Log($"{savePath}已存在, 跳过下载...");
-                    relatedTask?.SavePaths.Add(savePath);
-                    if (selectedPagesInfo.Count == 1 && Directory.Exists(p.aid))
-                    {
-                        Directory.Delete(p.aid, true);
-                    }
-                    return;
-                }
-                var pad = string.Empty.PadRight(clips.Count.ToString().Length, '0');
-                for (int i = 0; i < clips.Count; i++)
-                {
-                    var link = clips[i];
-                    videoPath = $"{p.aid}/{p.aid}.P{p.index}.{p.cid}.{i.ToString(pad)}.mp4";
-                    Log($"开始下载P{p.index}视频, 片段({(i + 1).ToString(pad)}/{clips.Count})...");
-                    await DownloadTrackAsync(link, videoPath, downloadConfig, video: true);
-                }
-                Log($"下载P{p.index}完毕");
-                Log("开始合并分段...");
-                var files = GetFiles(Path.GetDirectoryName(videoPath)!, ".mp4");
-                videoPath = $"{p.aid}/{p.aid}.P{p.index}.{p.cid}.mp4";
-                BBDownMuxer.MergeFLV(files, videoPath);
-                if (myOption.SkipMux) return;
-                Log($"开始混流视频{(subtitleInfo.Any() ? "和字幕" : "")}...");
-                if (myOption.AudioOnly)
-                    savePath = savePath[..^4] + ".m4a";
-                int code = BBDownMuxer.MuxAV(false, p.bvid, videoPath, "", audioMaterial, savePath,
-                    desc,
-                    title,
-                    p.ownerName ?? "",
-                    (pagesCount > 1 || (bangumi && !vInfo.IsBangumiEnd)) ? p.title : "",
-                    File.Exists(coverPath) ? coverPath : "",
-                    lang,
-                    subtitleInfo, myOption.AudioOnly, myOption.VideoOnly, p.points, p.pubTime, myOption.SimplyMux);
-                if (code != 0 || !File.Exists(savePath) || new FileInfo(savePath).Length == 0)
-                {
-                    LogError("合并失败"); return;
-                }
-                Log("清理临时文件...");
-                await Task.Delay(200);
-                if (parsedResult.VideoTracks.Count != 0) File.Delete(videoPath);
-                foreach (var s in subtitleInfo) File.Delete(s.path);
-                foreach (var a in audioMaterial) File.Delete(a.path);
-                if (p.points.Any()) File.Delete(Path.Combine(Path.GetDirectoryName(string.IsNullOrEmpty(videoPath) ? audioPath : videoPath)!, "chapters"));
-                if (selectedPagesInfo.Count == 1 || p.index == selectedPagesInfo.Last().index || p.aid != selectedPagesInfo.Last().aid)
-                    File.Delete(coverPath);
-                if (Directory.Exists(p.aid) && Directory.GetFiles(p.aid).Length == 0) Directory.Delete(p.aid, true);
-            }
-            else
-            {
-                if (myOption.DecryptDrm)
-                {
-                    LogError("此视频需要大会员登录才能获取完整DRM内容。");
-                    LogError("请先运行: BBDown login  或使用 --cookie 参数");
-                }
-                else
-                {
-                    LogError("解析此分P失败(建议--debug查看详细信息)");
-                }
-                if (parsedResult.WebJsonString.Length < 100)
-                {
-                    LogError(parsedResult.WebJsonString);
-                }
-                LogDebug("{0}", parsedResult.WebJsonString);
-            }
-
-            if (!string.IsNullOrWhiteSpace(savePath)) {
-                relatedTask?.SavePaths.Add(savePath);
-            }
-            break; // success, exit retry loop
-        }
-        catch (Exception ex)
-        {
-            retryCount++;
-            if (retryCount >= 3) throw;
-            LogError(ex.Message);
-            LogWarn("下载出现异常, 3秒后将进行自动重试...");
-            await Task.Delay(3000);
-        }
-        }
-    }
-
     internal static async Task DoWorkAsync(MyOption myOption)
     {
         try
@@ -793,235 +141,4 @@ partial class Program
         }
     }
 
-    private static List<Video> SortTracks(List<Video> videoTracks, Dictionary<string, int> dfnPriority, Dictionary<string, byte> encodingPriority, bool videoAscending)
-    {
-        // 编码优先：先按编码排序，再按清晰度排序；清晰度优先时使用 --dfn-priority 即可
-        return videoTracks
-            .OrderBy(v => encodingPriority.GetValueOrDefault(v.codecs, (byte)100))
-            .ThenBy(v => dfnPriority.GetValueOrDefault(v.dfn, 100))
-            .ThenByDescending(v => Convert.ToInt32(v.id))
-            .ThenBy(v => videoAscending ? v.bandwidth : -v.bandwidth)
-            .ToList();
-    }
-    
-    private static List<Audio> SortTracks(List<Audio> audioTracks, Dictionary<string, byte> encodingPriority, bool audioAscending)
-    {
-        return audioTracks
-            .OrderBy(a => encodingPriority.GetValueOrDefault(a.shortCodecs, (byte)100))
-            .ThenBy(a => audioAscending ? a.bandwidth : -a.bandwidth)
-            .ToList();
-    }
-
-    private static string FormatSavePath(string savePathFormat, string title, Video? videoTrack, Audio? audioTrack, Page p, int pagesCount, string apiType, long pubTime)
-    {
-        var result = savePathFormat.Replace('\\', '/');
-        var regex = InfoRegex();
-        foreach (Match m in regex.Matches(result).Cast<Match>())
-        {
-            var key = m.Groups[1].Value;
-
-            //解析自定义日期格式
-            var defaultDateFormat = "yyyy-MM-dd_HH-mm-ss";
-            string[] prefixes = ["publishDate:", "videoDate:"];
-            foreach (var prefix in prefixes)
-            {
-                if (key.StartsWith(prefix))
-                {
-                    defaultDateFormat = key[(key.IndexOf(':') + 1)..];
-                    key = prefix.Replace(":", "");
-                    break;
-                }
-            }
-
-            var v = key switch
-            {
-                "videoTitle" => GetValidFileName(title, filterSlash: true).Trim().TrimEnd('.').Trim(),
-                "pageNumber" => p.index.ToString(),
-                "pageNumberWithZero" => p.index.ToString().PadLeft(pagesCount.ToString().Length, '0'),
-                "pageTitle" => GetValidFileName(p.title, filterSlash: true).Trim().TrimEnd('.').Trim(),
-                "bvid" => p.bvid,
-                "aid" => p.aid,
-                "cid" => p.cid,
-                "ownerName" => p.ownerName == null ? "" : GetValidFileName(p.ownerName, filterSlash: true).Trim().TrimEnd('.').Trim(),
-                "ownerMid" => p.ownerMid ?? "",
-                "dfn" => videoTrack == null ? "" : videoTrack.dfn,
-                "res" => videoTrack == null ? "" : videoTrack.res,
-                "fps" => videoTrack == null ? "" : videoTrack.fps,
-                "videoCodecs" => videoTrack == null ? "" : videoTrack.codecs,
-                "videoBandwidth" => videoTrack == null ? "" : videoTrack.bandwidth.ToString(),
-                "audioCodecs" => audioTrack == null ? "" : audioTrack.codecs,
-                "audioBandwidth" => audioTrack == null ? "" : audioTrack.bandwidth.ToString(),
-                "publishDate" => FormatTimeStamp(pubTime, defaultDateFormat),
-                "videoDate" => FormatTimeStamp(p.pubTime, defaultDateFormat),
-                "apiType" => apiType,
-                _ => $"<{key}>"
-            };
-            result = result.Replace(m.Value, v);
-        }
-        if (!result.EndsWith(".mp4")) { result += ".mp4"; }
-        return result;
-    }
-
-    [GeneratedRegex("<([\\w:\\-.]+?)>")]
-    private static partial Regex InfoRegex();
-
-    private static async Task DecryptDrmAsync(ParsedResult parsed, string videoPath, string audioPath, MyOption myOption)
-    {
-        Log("检测到DRM加密，正在获取解密密钥...");
-
-        parsed.KeyHex = myOption.DrmKeyHex ?? "";
-        if (!string.IsNullOrEmpty(myOption.DrmKidHex))
-            parsed.KidHex = myOption.DrmKidHex;
-
-        if (!string.IsNullOrEmpty(parsed.KeyHex) && !string.IsNullOrEmpty(parsed.KidHex))
-        {
-            Log($"使用手动提供的密钥: KEY={parsed.KeyHex[..Math.Min(8, parsed.KeyHex.Length)]}...");
-        }
-        else
-        {
-            try
-            {
-            if (parsed.DrmTechType == 2)
-            {
-                if (!string.IsNullOrEmpty(parsed.PsshBase64))
-                {
-                    var wvd = !string.IsNullOrEmpty(myOption.WvdPath) && File.Exists(myOption.WvdPath)
-                        ? myOption.WvdPath
-                        : FindTool("device.wvd") ?? Path.Combine(AppContext.BaseDirectory, "device.wvd");
-                    if (File.Exists(wvd))
-                    {
-                        var keyResult = await DrmDecryptor.GetKeyWidevineAsync(parsed.PsshBase64, wvd);
-                        if (keyResult != null)
-                        {
-                            parsed.KeyHex = keyResult.Value.keyHex;
-                            parsed.KidHex = keyResult.Value.kid;
-                        }
-                    }
-                    else
-                    {
-                        LogWarn("Widevine DRM 需要 device.wvd 文件，请放置到程序目录");
-                    }
-                }
-            }
-            else
-            {
-                LogWarn("当前DRM类型不支持自动解密，请使用 --key --kid 手动提供密钥");
-            }
-        }
-        catch (Exception ex) { LogWarn($"自动密钥提取异常: {ex.Message}"); }
-
-        if (string.IsNullOrEmpty(parsed.KeyHex))
-        {
-            LogWarn("============================================");
-            LogWarn("自动密钥提取失败，文件将保持加密状态。");
-            LogWarn("");
-            LogWarn("解决方案：");
-            LogWarn("  1. 确保 device.wvd 文件放置在程序目录下");
-            LogWarn($"  2. 或手动指定: BBDown <url> --key <KEY_HEX> --kid {parsed.KidHex}");
-            LogWarn("============================================");
-            return;
-        }
-        }
-
-        Log($"密钥获取成功: KEY={parsed.KeyHex[..Math.Min(8, parsed.KeyHex.Length)]}...");
-
-        var mp4decrypt = !string.IsNullOrEmpty(myOption.Mp4decryptPath) && File.Exists(myOption.Mp4decryptPath)
-            ? myOption.Mp4decryptPath
-            : FindTool("mp4decrypt");
-        if (string.IsNullOrEmpty(mp4decrypt))
-        {
-            LogError("未找到 mp4decrypt，请安装 Bento4 或通过 --mp4decrypt-path 指定路径");
-            return;
-        }
-
-        if (!string.IsNullOrEmpty(videoPath) && File.Exists(videoPath))
-        {
-            Log("解密视频流...");
-            var tmpVideo = videoPath + ".dec";
-            await RunDecryptAsync(mp4decrypt, parsed.KidHex, parsed.KeyHex, videoPath, tmpVideo);
-            if (File.Exists(tmpVideo) && new FileInfo(tmpVideo).Length > 0)
-            {
-                File.Delete(videoPath);
-                File.Move(tmpVideo, videoPath);
-                Log("视频解密完成");
-            }
-        }
-
-        if (!string.IsNullOrEmpty(audioPath) && File.Exists(audioPath))
-        {
-            Log("解密音频流...");
-            var tmpAudio = audioPath + ".dec";
-            await RunDecryptAsync(mp4decrypt, parsed.KidHex, parsed.KeyHex, audioPath, tmpAudio);
-            if (File.Exists(tmpAudio) && new FileInfo(tmpAudio).Length > 0)
-            {
-                File.Delete(audioPath);
-                File.Move(tmpAudio, audioPath);
-                Log("音频解密完成");
-            }
-        }
-    }
-
-    private static async Task RunDecryptAsync(string mp4decrypt, string kid, string key, string input, string output)
-    {
-        var psi = new ProcessStartInfo
-        {
-            FileName = mp4decrypt,
-            Arguments = $"--key {kid}:{key} \"{input}\" \"{output}\"",
-            RedirectStandardOutput = false,
-            RedirectStandardError = true,
-            UseShellExecute = false,
-            CreateNoWindow = true,
-        };
-
-        using var proc = Process.Start(psi);
-        if (proc == null) return;
-        var stderrTask = proc.StandardError.ReadToEndAsync();
-        await proc.WaitForExitAsync();
-
-        if (proc.ExitCode != 0)
-        {
-            var err = await stderrTask;
-            LogError($"mp4decrypt failed: {err}");
-        }
-    }
-
-    private static string? FindTool(string name)
-    {
-        // 1. 优先搜索系统 PATH
-        var pathEnv = Environment.GetEnvironmentVariable("PATH");
-        var pathDirs = !string.IsNullOrEmpty(pathEnv)
-            ? pathEnv.Split(Path.PathSeparator, StringSplitOptions.RemoveEmptyEntries)
-            : Array.Empty<string>();
-
-        // 2. 然后搜索程序同目录及当前工作目录
-        var localDirs = new[] { AppContext.BaseDirectory, Environment.CurrentDirectory };
-
-        var allDirs = pathDirs.Concat(localDirs);
-
-        // Windows 下追加 .exe 后缀
-        var names = OperatingSystem.IsWindows()
-            ? new[] { name, name + ".exe" }
-            : new[] { name };
-
-        foreach (var dir in allDirs)
-        {
-            foreach (var n in names)
-            {
-                var full = Path.Combine(dir, n);
-                if (File.Exists(full)) return full;
-            }
-        }
-
-        // 3. Unix/macOS 常见安装路径回退
-        if (!OperatingSystem.IsWindows())
-        {
-            foreach (var dir in new[] { "/opt/homebrew/bin", "/usr/local/bin", "/usr/bin" })
-            {
-                var full = Path.Combine(dir, name);
-                if (File.Exists(full)) return full;
-            }
-        }
-
-        return null;
-    }
 }
